Cloudtask Agent fix
1,job run modification
Generate a script file for each job, windows is the run.cmd script, and linux is the run.sh script.
Copy job.cmd to the script file to prevent the command special characters in job.cmd from being transferred.

2, job stop Action modification
In the win32 platform, stop a job if it times out 5 seconds, then use the `taskkill` way to kill the work process tree.